# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RemoteGPU API
  description: RemoteGPU 远程GPU管理平台 API 文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@remotegpu.com
servers:
  - url: http://0.0.0.:8080/api/v1
    description: 开发环境
  - url: https://0.0.0.0:8080/api/v1
    description: 生产环境
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: 用户登录
      description: 使用用户名和密码登录系统
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 认证失败（业务错误码：2003=密码错误，2006=账号禁用）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_password:
                  summary: 密码错误
                  value:
                    code: 2003
                    msg: 密码错误
                    data: null
                account_disabled:
                  summary: 账号禁用
                  value:
                    code: 2006
                    msg: 账号已禁用
                    data: null
...
  /auth/profile:
    get:
      tags:
        - Auth
      summary: 获取用户资料
      description: 获取当前登录用户的个人资料
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/machines:
    get:
      tags:
        - Admin - Machines
      summary: 获取机器列表
      description: 获取所有机器的列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [available, allocated, maintenance]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Machine'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer

    post:
      tags:
        - Admin - Machines
      summary: 创建机器
      description: 添加新的机器到系统
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMachineRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/machines/import:
    post:
      tags:
        - Admin - Machines
      summary: 批量导入机器
      description: 批量导入机器信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                machines:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateMachineRequest'
      responses:
        '200':
          description: 导入成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success_count:
                    type: integer
                  failed_count:
                    type: integer
  /admin/machines/{id}/collect:
    post:
      tags:
        - Admin - Machines
      summary: 采集机器硬件信息
      description: 触发 Agent 采集并更新硬件信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 采集成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/machines/{id}/allocate:
    post:
      tags:
        - Admin - Machines
      summary: 分配机器
      description: 将机器分配给客户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocateRequest'
      responses:
        '200':
          description: 分配成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Allocation'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/machines/{id}/reclaim:
    post:
      tags:
        - Admin - Machines
      summary: 回收机器
      description: 从客户处回收机器
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReclaimRequest'
      responses:
        '200':
          description: 回收成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名
          example: admin
        password:
          type: string
          format: password
          description: 密码
          example: password123
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: 访问令牌
        refresh_token:
          type: string
          description: 刷新令牌
        expires_in:
          type: integer
          format: int64
          description: 过期时间（秒）
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, customer]
        created_at:
          type: string
          format: date-time
    AllocateRequest:
      type: object
      required:
        - customer_id
        - host_id
        - duration_months
      properties:
        customer_id:
          type: integer
          description: 客户ID
        host_id:
          type: string
          description: 主机ID
        duration_months:
          type: integer
          minimum: 1
          description: 分配时长（月）
        remark:
          type: string
          description: 备注
    ReclaimRequest:
      type: object
      properties:
        reason:
          type: string
          description: 回收原因
        force:
          type: boolean
          description: 是否强制回收
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误码（例如：2003=密码错误，2006=账号禁用）
          example: 2003
        msg:
          type: string
          description: 错误信息
          example: 密码错误
        data:
          nullable: true
          description: 可选错误数据
    Machine:
      type: object
      properties:
        id:
          type: string
        host_id:
          type: string
        hostname:
          type: string
        ip:
          type: string
        status:
          type: string
          enum: [available, allocated, maintenance]
        needs_collect:
          type: boolean
        cpu_cores:
          type: integer
        memory_gb:
          type: integer
        gpu_model:
          type: string
        gpu_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Allocation:
      type: object
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        machine_id:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
        remark:
          type: string
...
    CreateMachineRequest:
      type: object
      required:
        - name
        - region
        - ssh_port
        - ssh_username
      properties:
        name:
          type: string
        hostname:
          type: string
        region:
          type: string
        ip_address:
          type: string
        public_ip:
          type: string
        ssh_port:
          type: integer
        ssh_username:
          type: string
        ssh_password:
          type: string
        ssh_key:
          type: string
...